Section 19 - Pre switch to postgres

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {Message} from "../_models/message";
 import {Pagination} from "../_models/pagination";
 import {MessageService} from "../_services/message.service";
+import {ConfirmService} from "../_services/confirm.service";
 
 @Component({
   selector: 'app-messages',
@@ -13,10 +14,12 @@ export class MessagesComponent implements OnInit {
   pagination!: Pagination
   container = 'Unread'
   pageNumber = 1
-  pageSize = 2
+  pageSize = 5
   loading = false
 
-  constructor(private messageService: MessageService) { }
+  constructor(
+    private messageService: MessageService,
+    private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages()
@@ -33,9 +36,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    // We don't get anything back from deleted. We use splice to find the index and specify how many messages we want to delete
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages?.splice(this.messages?.findIndex(m => m.id === id), 1)
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
+      // Result is going to be true or false.
+      if (result) {
+        // We don't get anything back from deleted. We use splice to find the index and specify how many messages we want to delete
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages?.splice(this.messages?.findIndex(m => m.id === id), 1)
+        })
+      }
     })
   }
